+++ thoth-api/migrations/0.8.3/down.sql
@@ -1,1 +1,3 @@
 DROP TRIGGER publication_location_canonical_urls_check ON publication;
+
+DROP FUNCTION IF EXISTS publication_location_canonical_urls();