+++ thoth-api/src/model/affiliation/crud.rs
@@ -16,6 +16,7 @@ impl Crud for Affiliation {
     type OrderByEntity = AffiliationOrderBy;
     type FilterParameter1 = ();
     type FilterParameter2 = ();
+    type FilterParameter3 = ();
 
     fn pk(&self) -> Uuid {
         self.affiliation_id
@@ -31,7 +32,8 @@ impl Crud for Affiliation {
         parent_id_1: Option<Uuid>,
         parent_id_2: Option<Uuid>,
         _: Vec<Self::FilterParameter1>,
-        _: Option<Self::FilterParameter2>,
+        _: Vec<Self::FilterParameter2>,
+        _: Option<Self::FilterParameter3>,
     ) -> ThothResult<Vec<Affiliation>> {
         use crate::schema::affiliation::dsl::*;
         let mut connection = db.get().unwrap();
@@ -97,7 +99,8 @@ impl Crud for Affiliation {
         _: Option<String>,
         _: Vec<Uuid>,
         _: Vec<Self::FilterParameter1>,
-        _: Option<Self::FilterParameter2>,
+        _: Vec<Self::FilterParameter2>,
+        _: Option<Self::FilterParameter3>,
     ) -> ThothResult<i32> {
         use crate::schema::affiliation::dsl::*;
         let mut connection = db.get().unwrap();