+++ thoth-api/src/model/imprint/crud.rs
@@ -18,6 +18,7 @@ impl Crud for Imprint {
     type OrderByEntity = ImprintOrderBy;
     type FilterParameter1 = ();
     type FilterParameter2 = ();
+    type FilterParameter3 = ();
 
     fn pk(&self) -> Uuid {
         self.imprint_id
@@ -33,7 +34,8 @@ impl Crud for Imprint {
         parent_id_1: Option<Uuid>,
         _: Option<Uuid>,
         _: Vec<Self::FilterParameter1>,
-        _: Option<Self::FilterParameter2>,
+        _: Vec<Self::FilterParameter2>,
+        _: Option<Self::FilterParameter3>,
     ) -> ThothResult<Vec<Imprint>> {
         use crate::schema::imprint::dsl::*;
         let mut connection = db.get().unwrap();
@@ -89,7 +91,8 @@ impl Crud for Imprint {
         filter: Option<String>,
         publishers: Vec<Uuid>,
         _: Vec<Self::FilterParameter1>,
-        _: Option<Self::FilterParameter2>,
+        _: Vec<Self::FilterParameter2>,
+        _: Option<Self::FilterParameter3>,
     ) -> ThothResult<i32> {
         use crate::schema::imprint::dsl::*;
         let mut connection = db.get().unwrap();