+++ thoth-api/src/model/work_relation/crud.rs
@@ -17,6 +17,7 @@ impl Crud for WorkRelation {
     type OrderByEntity = WorkRelationOrderBy;
     type FilterParameter1 = RelationType;
     type FilterParameter2 = ();
+    type FilterParameter3 = ();
 
     fn pk(&self) -> Uuid {
         self.work_relation_id
@@ -32,7 +33,8 @@ impl Crud for WorkRelation {
         parent_id_1: Option<Uuid>,
         _: Option<Uuid>,
         relation_types: Vec<Self::FilterParameter1>,
-        _: Option<Self::FilterParameter2>,
+        _: Vec<Self::FilterParameter2>,
+        _: Option<Self::FilterParameter3>,
     ) -> ThothResult<Vec<WorkRelation>> {
         use crate::schema::work_relation::dsl::*;
         let mut connection = db.get().unwrap();
@@ -91,7 +93,8 @@ impl Crud for WorkRelation {
         _: Option<String>,
         _: Vec<Uuid>,
         relation_types: Vec<Self::FilterParameter1>,
-        _: Option<Self::FilterParameter2>,
+        _: Vec<Self::FilterParameter2>,
+        _: Option<Self::FilterParameter3>,
     ) -> ThothResult<i32> {
         use crate::schema::work_relation::dsl::*;
         let mut connection = db.get().unwrap();