+++ thoth-app/src/models/work/create_work_mutation.rs
@@ -82,6 +82,7 @@ const CREATE_WORK_MUTATION: &str = "
             imprintId
             createdAt
             updatedAt
+            updatedAtWithRelations
         }
     }
 ";